Makes the daemon list only its remote refs.
We only fetch objects which are anounced as remote. Currently, the tests
ain't working, because we don't have remote objects until we do the
first push. The problem is, the push functionality isn't implemented
yet.

I don't know whether the do_list should be reworked for the push
functionality. How do we determine what to push?