Implements ref listing depending on latest refs.
Now info about each peer interested in particular repo is stored in a
database. The info contains the latest ref the peer has about a
particular branch. When git asks us to list the refs, the relay collects
all latest peer refs and determines which is the most recent. It then
gives it to us so we could fetch the proper objects.